Added pic stitching script
Python Pic Stitching Script

Give the script a working directory (such as /var/www/html/users/"userID"), the number of images you wish to add, and the names of the image (1.jpeg, 2.jpeg, etc.). This script then stitches them together to a file (placed in the current directory) called target.jpeg, which is either created or written over by this script. Pictures do not need to be added multiple times. Once it is run through the script once, it will be in the target image.